Update import paths in a package.
Implemented using go/parser and golang.org/x/tools/astutil, able to
parse a directory edit import paths according to a map[string]string
and write back to necessary files.

Updated specification, I realized that editing qualified identifiers
wasn't necessary when moving or copying since the package defaults to
the package name defined in the package source, not the basename of the
directory. Another subcommand was introduced, `name`, that will handle
renaming a package and updating the qualified identifiers in a package.

Fixed an issue with `getPackage` not properly computing the absolute
path.