Recursively update import paths.
Fixed problem that when a package was copied it did not update the
import paths in the copied package recursively, copied child packages
would end up using the old import path.

Added recurse options on the `mv` and `cp` subcommand which recursively
update import paths in the subdirectories as well.